Trigger E_USER_DEPRECATED errors for deprecated methods
This is the recommended way to deprecate functionality.
If you need to actually keep using the old interface, but still
upgrading to a new version you could disable deprecated errors from being
issues with:

    error_reporting(E_ALL & ~E_USER_DEPRECATED);

Unfortunately these are not easy to test with the latest PHPUnit version, so
I have suppressed these errors in the tests.