Pass on $code and $previous to parent constructors on Swift Mailer exceptions
Before SwiftMailer 5.4.1, SwiftMailer did not conform to the PHP Exception class signature.
So we are setting the $code which is not private so it is set regardless of that.
In most cases this would work properly both before and after 5.4.1.
If you are using SwiftMailer < 5.4.1, you would not get the $previous exception set and you would
if you are using a newer version of SwiftMailer.