Support swiftmailer 6
This drops support of the swiftmailer 5 library in favour of swiftmailer
6.

The `Swift_Mime_Message` interface is no longer available in swiftmailer
6. Instead it is replaced with the concrete `Swift_Mime_SimpleMessage`
class which is functionally equivalent.

A `Swift_Transport_PostmarkTransport::ping()` method is added to meet
the newer `Swift_Transport` interface. This method is intended to 'Check
if this Transport mechanism is alive'. Given we are using a
transactional API, we can just return true for the ping (as we have no
persistent connection to manage).

`newInstance()` is also removed from `Swift_Message` and `Swift_Mailer`
in swiftmailer 6. Instead these instances are just initialised via normal
object construction.